Store results of file-accessing functions in the correct data types
The `::read` and `::lseek` functions return a `ssize_t` and `off_t` respectively, which are defined as 64-bit types on a 64-bit machine. Storing them in the 32-bit `int` type has led to signed-integer overflow on the large files created by `buck`. This diff ensures that the calls to these `stdlib` functions match the signature specified in their `man` pages.

I've also vetted the rest of the `File` functions to ensure that all of them store the results of their `stdlib` calls in the correct data type.

Based on Facebook D26484212

+++ runtime/file-darwin.cpp
@@ -41,7 +41,7 @@ int File::open(const char* path, int flags, int mode) {
 }
 
 ssize_t File::read(int fd, void* buffer, size_t count) {
-  int result;
+  ssize_t result;
   do {
     result = ::read(fd, buffer, count);
   } while (result == -1 && errno == EINTR);
@@ -54,7 +54,7 @@ int File::setNoInheritable(int fd) {
 }
 
 int64_t File::seek(int fd, int64_t offset, int whence) {
-  int result = ::lseek(fd, offset, whence);
+  off_t result = ::lseek(fd, offset, whence);
   return result < 0 ? -errno : result;
 }
